Revert "Fixed dimension of Fcnt header field according to specifications"
This reverts commit 3c8221a37fb8a8bd22ca8074186856fb9c159548.

+++ model/end-device-lorawan-mac.h
@@ -491,7 +491,7 @@ private:
    */
   LorawanMacHeader::MType m_mType;
 
-  uint16_t m_currentFCnt;
+  uint8_t m_currentFCnt;
 };
